% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R
\docType{data}
\name{airquality}
\alias{airquality}
\title{airquality Dataset}
\format{
A data frame with 153 rows and 6 columns:
\describe{
\item{\code{Ozone}}{numeric.  Mean ozone in parts per billion from 1300 to 1500 hours at Roosevelt Island.}
\item{\code{Solar.R}}{numeric. Solar radiation in Langleys in the frequency band 4000–7700 Angstroms from 0800 to 1200 hours at Central Park.}
\item{\code{Wind}}{numeric. Average wind speed in miles per hour at 0700 and 1000 hours at LaGuardia Airport.}
\item{\code{Temp}}{numeric. Charles River dummy variable (= 1 if tract bounds river; 0 otherwise).}
\item{\code{Month}}{numeric. Month(1-12).}
\item{\code{Day}}{numeric. Day of month (1--31).}
}
}
\source{
base R
}
\usage{
airquality
}
\description{
A dataframe that is derived from information collected about the daily air quality measurements in New York,
May to September 1973.
}
\keyword{datasets}
