% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit}
\alias{audit}
\title{Audit of a dataframe}
\usage{
audit(
  data,
  catCut = 0.05,
  levels_numeric = 4,
  levels_categorical = 5,
  row_percent = 35,
  col_percent = 20
)
}
\arguments{
\item{data}{a dataframe}

\item{catCut}{minimum percentage of values in a category in order to not be an outlier}

\item{levels_numeric}{minimum number of unique values that a numeric variable should have}

\item{levels_categorical}{maximum number of categories that a categorical variable should have}

\item{row_percent}{the upper bound percent value for missing values in a row}

\item{col_percent}{the upper bound percent value for missing values in a column}
}
\value{
a message in the console that list an analysis on the missing values of the data, the categorical and numeric variables and the outliers
based on these datatypes
}
\description{
This \code{audit} function is used to assess a data frame. It will return the percentage of values that are missing, where the
percentage is dependent on user input. It also returns information on the outlier of the categorcial and nuemric data.
}
\details{
Once given a dataframe it will return an audit that describes the data frame.
}
\examples{
audit(mtcars)
audit(airquality)

}
