% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R, R/mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{airquality Dataset}
\format{
A data frame with 153 rows and 6 columns:
\describe{
\item{\code{Ozone}}{numeric.  Mean ozone in parts per billion from 1300 to 1500 hours at Roosevelt Island.}
\item{\code{Solar.R}}{numeric. Solar radiation in Langleys in the frequency band 4000–7700 Angstroms from 0800 to 1200 hours at Central Park.}
\item{\code{Wind}}{numeric. Average wind speed in miles per hour at 0700 and 1000 hours at LaGuardia Airport.}
\item{\code{Temp}}{numeric. Charles River dummy variable (= 1 if tract bounds river; 0 otherwise).}
\item{\code{Month}}{numeric. Month(1-12).}
\item{\code{Day}}{numeric. Day of month (1--31).}
}

A data frame with 153 rows and 6 columns:
\describe{
\item{\code{mpg}}{numeric.  Miles/(US) gallon.}
\item{\code{cyl}}{numeric. Number of cylinders.}
\item{\code{disp}}{numeric. Displacement (cu.in.).}
\item{\code{hp}}{numeric. Gross horsepower.}
\item{\code{drat}}{numeric. Rear axle ratio.}
\item{\code{wt}}{numeric. Weight (1000 lbs).}
\item{\code{qsec}}{numeric. 1/4 mile time.}
\item{\code{vs}}{numeric. Engine (0 = V-shaped, 1 = straight)}
\item{\code{am}}{numeric. 	Transmission (0 = automatic, 1 = manual)}
\item{\code{gear}}{numeric. Number of forward gears}
}
}
\source{
base R

base R
}
\usage{
mtcars

mtcars
}
\description{
A dataframe that is derived from information collected about the daily air quality measurements in New York,
May to September 1973.

A dataframe that is derived from the 1974 Motor Trend US magazine, and comprises fuel consumption and 10 aspects of automobile design and performance for 32 automobiles (1973–74 models)
}
\keyword{datasets}
